scripts/python/convert_all_heic_to_jpg.requirements.txt
cffi==1.15.1
imagecodecs==2023.3.16
numpy==1.24.3
Pillow==9.5.0
pycparser==2.21
pyheif==0.7.1
pyheif-pillow-opener==0.2.0
